@@ -17,117 +16,76 @@
   },
   "main": "index.js",
   "license": "Apache-2.0",
+  "engines": {
+    "node": ">=18.15.0"
+  },
   "dependencies": {
-    "chalk": "^4.1.0",
-    "lit-element": "^2.4.0"
+    "chalk": "^5.3.0",
+    "lit": "^2.7.6"
   },
   "peerDependencies": {
-    "@alaskaairux/design-tokens": "^3.9.0",
-    "@alaskaairux/icons": "^4.13.0",
-    "@alaskaairux/webcorestylesheets": "^3.7.3",
-    "@webcomponents/webcomponentsjs": "^2.5.0",
-    "focus-visible": "^5.2.0"
+    "@alaskaairux/design-tokens": "^3.14.0",
+    "@alaskaairux/icons": "^4.28.0",
+    "@aurodesignsystem/webcorestylesheets": "^4.6.1"
   },
   "devDependencies": {
-    "@alaskaairux/design-tokens": "3.9.0",
-    "@alaskaairux/icons": "4.13.0",
-    "@alaskaairux/webcorestylesheets": "3.7.3",
-    "@babel/core": "^7.13.14",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.13.10",
-    "@babel/preset-env": "^7.13.12",
-    "@babel/runtime": "^7.13.10",
-    "@commitlint/cli": "^12.1.1",
-    "@commitlint/config-conventional": "^12.1.1",
-    "@open-wc/testing": "^2.5.32",
-    "@open-wc/testing-karma": "^4.0.9",
-    "@rollup/plugin-alias": "^3.1.9",
-    "@rollup/plugin-babel": "^5.3.1",
-    "@rollup/plugin-commonjs": "^18.0.0",
-    "@rollup/plugin-node-resolve": "^11.2.1",
-    "@semantic-release/changelog": "^5.0.1",
-    "@semantic-release/git": "^9.0.0",
-    "@semantic-release/npm": "^7.1.0",
-    "@web/dev-server": "^0.1.31",
-    "@webcomponents/webcomponentsjs": "^2.5.0",
-    "autoprefixer": "^10.2.5",
-    "babel-loader": "^8.2.5",
+    "@alaskaairux/design-tokens": "^3.14.0",
+    "@alaskaairux/icons": "^4.28.0",
+    "@aurodesignsystem/eslint-config": "^1.3.0",
+    "@aurodesignsystem/webcorestylesheets": "^4.6.1",
+    "@commitlint/cli": "^17.6.6",
+    "@commitlint/config-conventional": "^17.6.6",
+    "@open-wc/testing": "^3.2.0",
+    "@rollup/plugin-node-resolve": "^15.1.0",
+    "@semantic-release/changelog": "^6.0.3",
+    "@semantic-release/git": "^10.0.1",
+    "@semantic-release/npm": "^10.0.4",
+    "@web/dev-server": "^0.3.0",
+    "@web/test-runner": "^0.17.0",
+    "autoprefixer": "^10.4.14",
     "compression": "^1.7.4",
     "concat": "^1.0.3",
+    "concurrently": "^8.2.0",
     "copyfiles": "^2.4.1",
-    "core-js": "^3.10.0",
-    "eslint": "^7.23.0",
-    "eslint-plugin-lit": "^1.3.0",
-    "focus-visible": "^5.2.0",
-    "husky": "^4.3.8",
+    "core-js": "^3.31.1",
+    "eslint": "^8.44.0",
+    "eslint-plugin-jsdoc": "^46.4.3",
+    "husky": "^8.0.0",
     "lodash": "^4.17.21",
-    "marked": "^2.0.1",
-    "nodemon": "^2.0.16",
+    "markdown-magic": "^2.6.1",
+    "nodemon": "^3.0.1",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.2.9",
-    "postcss-custom-properties": "^11.0.0",
-    "postcss-discard-comments": "^4.0.2",
-    "postcss-remove-rules": "^1.0.0",
+    "postcss": "^8.4.26",
+    "postcss-discard-comments": "^6.0.0",
     "postcss-selector-replace": "^1.0.2",
-    "prismjs": "^1.23.0",
-    "rollup": "^2.44.0",
-    "rollup-plugin-minify-html-literals": "^1.2.6",
-    "rollup-plugin-serve": "^1.1.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "semantic-release": "^17.4.2",
-    "sinon": "^10.0.0",
-    "stylelint": "^13.12.0",
-    "stylelint-config-standard": "^21.0.0",
-    "wc-sass-render": "^1.2.3",
-    "web-component-analyzer": "^1.1.6",
-    "webpack-merge": "^5.7.3",
+    "rollup": "^3.26.2",
+    "rollup-plugin-serve": "^2.0.2",
+    "sass": "^1.63.6",
+    "semantic-release": "^21.0.7",
+    "stylelint": "^15.10.1",
+    "stylelint-config-idiomatic-order": "^9.0.0",
+    "stylelint-config-standard": "^34.0.0",
+    "stylelint-config-standard-scss": "^10.0.0",
+    "stylelint-order": "^6.0.3",
+    "stylelint-scss": "^5.0.1",
+    "typescript": "^5.1.6",
+    "wc-sass-render": "^1.4.0",
+    "web-component-analyzer": "^1.1.7",
     "whatwg-fetch": "^3.6.2",
-    "yaml-lint": "^1.2.4"
+    "yaml-lint": "^1.7.0"
   },
   "browserslist": [
     "last 2 Chrome versions",
     "last 2 iOS major versions",
-    "ie 11",
     "last 2 Firefox versions",
     "last 2 Edge versions",
     "last 2 Safari major versions"
   ],
-  "scripts": {
-    "build": "npm run ciBuild",
-    "apiBuild": "wca analyze 'src/auro-datetime.js' --outFiles docs/api.md",
-    "bundler": "rollup -c",
-    "bundle:test": "rollup -c -w",
-    "postinstall": "node packageScripts/postinstall.js",
-    "cssLint": "stylelint \"./src/*.css\"",
-    "dev": "npm run sassBuild:watch",
-    "distJS": "copyfiles -u 1 -V './src/**/*.js' ./dist",
-    "sassRender": "sass-render src/*.css -t ./scripts/staticStyles-template.js",
-    "sweep": "rm -rf ./temp ./demo/css ./demo/fonts ./dist ./src/tokens | rm ./src/*.css ./src/*-css.js",
-    "esLint": "./node_modules/.bin/eslint src/auro-*.js",
-    "lint": "npm-run-all cssLint esLint",
-    "test": "karma start --coverage",
-    "test:watch": "karma start --auto-watch=true --single-run=false",
-    "test:ci": "npm-run-all test lint",
-    "ciBuild": "npm-run-all sassBuild sassRender cssLint distJS bundler postinstall apiBuild",
-    "sassBuild": "npm-run-all sassBuild:demo sassBuild:component postCss:component sassRender",
-    "sassBuild:dev": "npm-run-all sassBuild:demo sassBuild:component postCss:component sassRender",
-    "sassBuild:demo": "node-sass ./demo/sass/ --output ./demo/css/",
-    "sassBuild:component": "node-sass ./src/ --output ./src/",
-    "sassBuild:watch": "nodemon -e scss --watch src --exec npm run sassBuild:dev",
-    "postCss:component": "node ./scripts/postCss.js",
-    "serve": "web-dev-server --open demo/ --node-resolve --watch",
-    "demo:rmBuild": "rm -rf ./build",
-    "demo:newBuild": "mkdir ./build && mkdir ./build/css && mkdir ./build/demo",
-    "demo:copyDemo": "copyfiles -u 1 -V './demo/demo.md' ./build/demo",
-    "demo:copyIndex": "copyfiles -u 1 -V './demo/index.html' ./build",
-    "demo:updateIndex": "node ./scripts/prepForBuild",
-    "demo:build": "npm-run-all ciBuild demo:rmBuild demo:newBuild demo:copyIndex demo:copyDemo demo:updateIndex"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "npm-run-all test lint bundler postinstall",
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
-    }
+  "nodemonConfig": {
+    "ignore": [
+      "*-css.js",
+      "*.css"
+    ]
   },
   "release": {
     "branches": [
