@@ -1,6 +1,8 @@
 'use strict';
 
-const chalk = require('chalk');
+import chalk from 'chalk';
+import { createRequire } from 'node:module';
+const require = createRequire(import.meta.url);
 const pjson = require('../package.json');
 
 console.log(chalk.hex('#f26135')(`
@@ -19,7 +21,7 @@ console.log(chalk.hex('#f26135')(`
 ╭ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ──────────────────────────────╮
 
         Thanks for installing the latest version
-                of `) + chalk.hex('#ffd200').bold(`auro-datetime v${pjson.version}.`) + chalk.hex('#f26135')(`
+                of `) + chalk.hex('#ffd200').bold(`auro-ddatetime v${pjson.version}.`) + chalk.hex('#f26135')(`
 
 ╰─────────────────────────────── ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─╯
 `)
