@@ -1,2 +1,2 @@
-import {css} from 'lit-element';
+import {css} from 'lit';
 export default css`<% content %>`;
